Code Clean up, Pokemon Addtitions, Abilities Added
cleaned up alot of the game manager code, splitting it into smaller more organized sections.  sprite atlas switched over to static methods. added pokemon from 26-36. added in alot of abilities as well as altering the functions within to be more resourceful to a wider range of abilties.

+++ Pokemon 2D Test Build/Assets/Scripts/Data/Abilitys/List/Overgrown.cs
@@ -17,7 +17,7 @@ public class Overgrown : AbilityBase
             return base.BoostACertainTypeInAPinch(attackingPokemon, attackType);
         }
 
-        if (attackingPokemon.currentHitPoints / attackingPokemon.maxHitPoints <= HpRequiredToActivatePinch)
+        if (((float)attackingPokemon.currentHitPoints / (float)attackingPokemon.maxHitPoints) <= HpRequiredToActivatePinch)
         {
             return 1.5f;
         }
